I somehow deleted the tests module __init__ without knowing it...
This commit fixes a lot of syntax errors and generally makes the repo work again
I'm getting too tired atm so this won't be very explanitory or not. But all tests
are now passing, and there shouldn't be any more future errors popping up, at least
any time soon with the exception of syntax errors because I'm a dumb-dumb.

This commit also adds a few more files to the tests library
	* encode.py for testing the encode method
	* compliments.py for testing the compliment_* methods
	* translations.py for testing the translate_* methods